Fix a defficiency in the RPN interface.
The RPN variable z-axis will be ordered as found in the file, which might not
be the expected monotic values.

For instance, if the model is outputting surface levels, and separately
outputting 3D fields, the surface level will appear first, followed by the
N-1 other levels from top to (near) bottom.