Implemented some basic support for masks.
If there are two records with identical headers, except one is a 1-bit integer,
then assume that it is a mask to apply to the other record.

An example of where this is used is in the RIOPS data (which is also an example
of where the X grid is needed).