Remove the old "rpn" module.
This was written from scratch, by reverse-engineering some sample RPN (FSTD)
files.  However, it doesn't not cover all possible cases and is fragile.

The "fstd" module gives the same functionality, but links to librmn to get
full compatibility with all FSTD files with a fraction of the code.

This also removes support for GEM "BMF" files, but I don't know of anyone who
actually used that interface anyway.