Add station names to netCDF output for timeseries data.
Note:  Kind of a dog's breakfast right now.  For instance, don't know why I
need to subtract 128 when decoding the bytes of the station names.

This whole class still needs a lot of work, but at least it's decoding the 2
sample files I have at my disposal.

Still need to figure out how to attach proper vertical coordinates.
Data will either be on thermodynamic / momentum levels, but I don't see how
that info is encoded.  Maybe just have to "know" this detail from the nomvar?

Ugh.  Not a fan of this FSTD variant at the moment.