Add support for scale_factor and add_offset.
This will allow unit conversion to be handled by the netCDF client when the
file is read.

For example, to convert the GEM model GZ field from decametres to metres, you
could create a metadata file with the contents:

[GZ]
standard_name = geopotential_height
scale_factor = 10
units = m

and then pass this extra file through the --metadata-file argument on the
command-line for fstd2nc.

Internally, the values will be stored in the original units (dam) as bitwise
copies of the original FST values, but when reading the data it should be
converted on-the-fly by the client (assuming a competent netCDF utility).

--

Also, auto-detect numerical values in the metadata files.