Check if the shape of lat/lon records matches the shape of the variable.
If not, complain to the user and then ignore the coordinate information.

This shouldn't normally happen for valid RPN files (I think?!).  The only way
I was able to trigger it is if I have multiple input files which have different
size of grids, but with identical ig1,ig2,ig3,ig4 values.

This fix doesn't handle the case where there are multiple grids with identical
*shapes* as well.  In that case, the output will just have the wrong grid for
some variables.