Implement gen_xyll function for all GridMaps.
This removes the need to check for the grid type to know which function to call.

xaxis and yaxis will only be used when they are required for the grid.

Also, check if the xaxis or yaxis have already been created.  For instance, on
a staggered horizontal grid (like with GEM UT1/VT1), they will appear multiple
times in different x/y combinations.