Prune unused axes from the data.
This was done to remove the extra 'time' axis when converting time series data.

Note: After doing this, a bug was found (maybe in python-netcdf4?) where using
an unlimited dimension somewhere besides the leftmost (outer) axis causes a
crash when writing the netcdf data (IndexError: tuple index out of range).

This happens in series data, where the time axis occurs after the station axis.

The bug wasn't triggered before, because the series data had an extra (unused)
time axis, which forced the axes to be renamed "time1" and "time2", which in
turn caused the unlimited dimension case to not be triggered.

As a workaround for now, don't try to set an unlimited dimension.