Check for wrapped longitude in LatLon.
Can happen for global unrotated 'E' grids.  In some cases, gdll sets the last
longitude to ~0 instead of 360.  This may cause problems in programs that
expect longitudes to be monotonically increasing.

This was already handled in the ezqkdef case of xycoords, but that code has
been recently superceded by the GridMap class.