Implement a to_pygeode() method to generate a pygeode Dataset.
Requires pygeode >= 1.2.0.

Note: There is currently a problem when there is an hour other than 0 in the
time axis.  The problem is that xarray encodes the time units with a character
'T' separating the date from the time, whereas pygeode assumes the separator
is always a space.  This should be handled in the next pygeode release (1.2.2).

See also: pygeode-rpn, another package for reading RPN files into pygeode.
          (https://github.com/neishm/pygeode-rpn)