Allow the user to pass in a dictionary-like object for caching the record headers from the files.
This cache can be used in subsequent fstd2nc.Buffer initializations to avoid
re-scanning files that are already known, especially useful when working on a
large number of files that will be used in multiple scripts.

It's up to the user to save this information somehow.

Note: This only caches the headers, not the data.