Refactor the code.
Two major changes are done here:

1) The mixins no longer use iterators to process the variables.  Instead, a
   list of variables is modified in-place with the relevant changes.
2) The structure of the variables is changed so that the axes/coordinates
   are stored directly in the structure, instead of matching by name / value
   combos.

With (1) and (2), it's now possible to remove all coordinate / axis info from
the list of variables, so the list only contains the data variables from the
FSTD file(s).

This will make is easier later on to implement a reverse process, where the
data structures are re-encoded into FSTD records.