Add experimental --fstd-compat option to make the netCDF file dual-function as an FSTD file.
This allows both the netCDF and RPN utilities to read the output file.  I'm
not entirely sure why someone would want to do this.  This is just a
demonstration that it's *technically* possible.

Compression is not available in this configuration, since netCDF4 and FSTD
do not share a common compression algorithm for their data.  Most data will
be encoded as 32-bit IEEE floats in big-endian order, which is what netCDF4
and FSTD can both easily read.

The implementation is a bit dodgy, and needs more testing, so it won't be
documented yet in command-line arguments.  It may also be removed entirely
if it's too impractical to maintain.