Add an option to only convert data encoded on thermodynamic levels.
Copied / pasted the logic from the diagnostic-level handling, modified as
appropriate.  Note: may not be the fastest approach, since it has to loop over
each momentum level individually and then check the whole table to see where
there's a match.

The only use case I can think of is when GZ is needed in conjunction with
thermodynamic fields (including tracers), in which case the combination of
--thermodynamic-levels and --ignore-diag-level should give a consistent vertical
coordinate.

Closes #29.