The --fstd-compat option should be workable enough for early testing.
Still experimental - depends on assumptions about where in the netCDF file the
data chunks are being written.  It's possible these assumptions will break in
future updates to the netCDF library.  This method of encoding definitely needs
to be refined.

Also may break some netCDF utilities that assume the netCDF4/HDF5 signature
block is at the start of the file (such as GDAL < 3.2.1).  The block had to be
relocated to offset 0x8000 to allow the file to start with the FSTD signature.