Some support for '#' grids.
Introduced a new _chunk_type object to handle fields that are partitioned
along i,j directions.  Added support for this object in netcdf and compat
mixins, and in the fstdump utility.

I don't have many examples of this type of grid, so there may be more work
needed for full support.