Change all_params function to take a Python file object as input, instead of a librmn file unit number.
Modified the code to extract the parameters directly from the file instead of
scraping them from the internal structures loaded by librmn.

The old approach depended on the inner workings of librmn, which could be
subject to change.  It also depended on the machine being little endian.

Now it only depends on the file structure itself, which is much less likely
to undergo change.