Update compat mixin to work with latest refactor.
No longer have access to internal structures of librmn to get directory pages,
so extract that information from fstprm instead.

Also, pad out the last page to 256 entries, since that seems to be the
convention of all the sample FSTD files, and is an assumption used in
all_params to read the headers back in.