Fix bug in series mixin, introduced by recent optimizations to date decoding.
The vectorized stamp2datetime is now returning a numpy datetime64 object, which
does not have an 'hour' attribute.

Create a separate routine for decoding single dates to datetime.datetime
objects, so the series mixin can get the information it needs.

Rename vectorized 'stamp2datetime' to 'stamp2datetim64' to be more clear about
the dtype being returned.