Add a --pandas option for convenvience during testing/debugging.
Allows the _makevars_pandas code path to be easily triggered on single-file
tests that would normally use _makevars_slow.

Note: _makevars_slow may be slower for large number of files, but for a single
file it turns out to be slower to import pandas than to just use this routine.