Remove leadtime and reftime from the "coordinates" attribute of variables.
The leadtime and reftime are still in the output, just not designated as
coordinates.

This addresses an issue with some netCDF viewers which have difficulty
interpreting these fields as coordinates.

I've added this fix since it seems that several projects are already adapting
the workaround of passing "--exclude=leadtime,reftime" to avoid this problem.

Closes #27.