Remove most occurrences of _fstluk method in the mixins.
This was the old method of reading the data, but required the use of thread
locks to avoid race conditions within librmn.

This is replaced by the _decode method, which avoids the thread-unsafe parts
of librmn.

There is still one occurrence of _fstluk, used locally within the compat mixin
(but not needed outside the scope of that particular code).

Closes #44.