Start removing the use of special metadata file (_meta_funit) for accessing coordinate records.
So far, implemented changes for series, vcoords, and xycoords mixins.

Some extra code was needed in xycoords to manually decode 'U' grids instead of
using librmn's functions.  This is because the ezqkdef function requires
side-channel access to the FST file to look up an additional record, which is
what we're trying to avoid here.

For similar reasons, some extra code was added to handle 'Z' grids without
using ezqkdef.

Note: these changes break the functionality of the gridhacks mixins.  They will
need to be updated to pass the modified grid info directly through the table.