A few speedups for some of the mixins.
filter: only generate unicode version of string columns if the column name is
mentioned in a filter.

series: early return if no series data was found.

vcoords: limit the number of vertical records to process, to speed up case
where many input files are being scanned.

extern: use np.sum instead of built-in sum in part of the fusing logic.