Fix a bug in _makevars_slow when trying to create an 'accum' axis on certain variables that don't actually need an 'accum' axis.
If some (but not all) of the 'accum' values were masked out for that variable,
it was having trouble organizing the record ids.

There was an inconsistency where the record ids had a reduced dimensionality
along the 'accum' axis (with length 1), but the logic for matching records to
a location in the record_id array was trying to use an n+1 index
for 'NaN' values.