Merge in updates from old 'rsf' branch.
This code was branched from last year, so there were a lot of merge conflicts
to resolve to bring it up-to-date.   Hopefully the changes got merged properly.

This should add some support for reading RSF files through the fstd Buffer
interface, but will only support the old-style (fstd98) metadata.

Closes #69.

A separate Buffer-type object interface would have to be created once the
fstd24 metadata structure is finalized and published.