Fix RSF decoder for more recent alpha versions of the file structure.
The size of the record header has changed, and may continue to change, so
remove hard-coded assumptions about that.

Some updates to extended metadata decoder to be more robust in the face of
these header size changes.  Also, the header size info seems to be encoded in
different ways for different alpha versions, so try to look in as many known
locations as possible to get this information.