Fix a bug in the new _split_meta internal routine.
It needs to exclude the aux keys from the header data.  Otherwise, header
width becomes 20 words when decoding metadata from fstd98 data segment, and
since that's more than 18 it detects as fst24/RSF formatted metadata in other
parts of the code and tries unsuccessfully to decode the non-existent RSF
attributes.