Fix a bug when ip2=0, ip3=0, and ip1 includes zero and non-zero values.
This was tripping up the masking in the 'accum' mixin, which was setting an
accum value for ip1=0 and masking out the other ones.  This partially-masked
axis was tripping up the indexing in makevars_slow, which only saw the non-nan
values in the 'accum' axis array.

The fix is to append a single 'nan' entry to the axis when it is partially
masked, similar to what is happening in the makevars_pandas version.

This fix is related to 0ed69002d67ad9b81d12b7ae70cde00892bab5e0, in which other
problems with accum axis masking were resolved.