Fix bug in endianness detection for files created with librmn 20.1-alpha.
One of the "pad" bytes in the record headers was used for determining the
endian of the data in fstd2nc, but that byte has recently been repurposed to
hold the number of bits of the data type(?).  This was causing fstd2nc to
erroneously treat the FST98 headers as little-endian encoded, giving wrong
values for the metadata.

Fixed it for now, by assuming that all FST98 data is big-endian, and all FST24
data is little-endian.

Endianness is encoded in a separate part of the RSF files, so in theory that
information is readily available for FST24, however it's not in the record
header section, so the decoding routines would have to be re-written to get
some extra context (not doing that for a bugfix release).