Fix bug in to_fstpy, which was causing OOM errors when constructing the 'd' column of the headers.
Passing a list of dask arrays to fill in the values seems to trigger a weird
transformation of 2D dask arrays into 2D array of dengenerate dask objects,
which is extremely ineffecient and weird.

By doing an explicit copy of each dask array in a loop, this weird behaviour is
avoided.

I don't recall encountered a problem with this code in the past, so maybe it's
a changed in behaviour in recent versions of numpy?