Some precision fixes for re-encoding grid information.
Recent versions of xarray load 'float' netCDF attributes as numpy.float32
values, and math on those values is done in single precision.  This was
affecting the code that looked for matching grid parameters for encoding back
to FST format.

This should address one of the problems encountered in #105.

Also force integer attributes to be 'int' (in case they might get put into
numpy.int32 values).  The rpnpy functions can be picky about data types, don't
want it to crash unexpectedly if the data type  for grid parameters is
unexpected.